refactor: 重构 1. compontents/article下文件移动至post文件夹下 2.代码格式优化
