--- src/components/home/card/signin/index.module.less
@@ -1,26 +1,0 @@
-
-@import '@/styles/variables.less';
-
-.card {
-  border: none;
-
-  .content {
-    .normal-flex(row, space-between, center);
-  }
-
-  .left {
-    .title {
-      color: var(--color-neutral-8);
-      font-size: 16px;
-      font-weight: 600;
-      line-height: 24px;
-    }
-
-    .desc {
-      color: var(--color-neutral-6);
-      font-size: 12px;
-      font-weight: 400;
-      line-height: 24px;
-    }
-  }
-}
