+++ src/components/home/card/welocome/index.tsx
@@ -6,18 +6,18 @@ interface Props {
 }
 
 const getTimeState = () => {
-  // 获取当前时间小时
+  // 获取当前时间小时 0-23
   const hours = new Date().getHours()
   // 设置默认文字
   let text = ''
   // 判断当前时间段
-  if (hours >= 0 && hours <= 10)
+  if (hours > 6 && hours <= 10)
     text = '早上好！'
-  else if (hours > 10 && hours <= 14)
+  else if (hours <= 14)
     text = '中午好！'
-  else if (hours > 14 && hours <= 18)
+  else if (hours <= 18)
     text = '下午好！'
-  else if (hours > 18 && hours <= 24)
+  else
     text = '晚上好！'
 
   return text
@@ -26,17 +26,17 @@ const getTimeState = () => {
 const Welcome: React.FC<Props> = (props: Props) => {
   const { slogan } = props
   return (
-        <div className={styles.myCard}>
-            <div>
-                <div className={styles.left}>
-                    <span>{getTimeState()}</span>
-                    <div>{slogan ?? '点亮在社区的每一天'}</div>
-                </div>
-                <button className={styles.right}>
-                    <span>去签到</span>
-                </button>
-            </div>
+    <div className={styles.myCard}>
+      <div>
+        <div className={styles.left}>
+          <span>{getTimeState()}</span>
+          <div>{slogan ?? '点亮在社区的每一天'}</div>
         </div>
+        <button className={styles.right}>
+          <span>去签到</span>
+        </button>
+      </div>
+    </div>
   )
 }
 Welcome.displayName = 'Welcome'
