+++ src/components/post/column/index.module.less
