+++ src/styles/variables.less
@@ -17,7 +17,7 @@
 
 .post-wrapper {
   width: @postWith;
-  margin: 20px auto 0;
+  margin: 80px auto 0;
   position: relative;
 }
